+++ blade-graphics/src/gles/mod.rs
@@ -479,8 +479,6 @@ impl Context {
             compute: self.capabilities.contains(Capabilities::BUFFER_STORAGE),
             max_compute_shared_memory_size: self.limits.max_compute_shared_memory_size,
             subgroup_size: 0,
-            min_subgroup_size: 0,
-            max_subgroup_size: 0,
             indirect_draw: false,
             binding_array: false,
             ray_query: crate::ShaderVisibility::empty(),